refactor: move the computation to tc.common
